Change Fixed Costs from type int to float

+++ powergenome/nrelatb.py
@@ -523,10 +523,10 @@ def atb_fixed_var_om_existing(
         ] = mod_results["Fixed_OM_Cost_per_MWyr"].mean()
     mod_results.loc[:, "Fixed_OM_Cost_per_MWyr"] = mod_results.loc[
         :, "Fixed_OM_Cost_per_MWyr"
-    ].astype(int)
+    ]
     mod_results.loc[:, "Fixed_OM_Cost_per_MWhyr"] = mod_results.loc[
         :, "Fixed_OM_Cost_per_MWhyr"
-    ].astype(int)
+    ]
     mod_results.loc[:, "Var_OM_Cost_per_MWh"] = mod_results.loc[
         :, "Var_OM_Cost_per_MWh"
     ]
